Remove `ThreadStateLogic` interface
