+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/logic/channel/thread/internal/ThreadStateLogic.kt
@@ -16,34 +16,43 @@
 
 package io.getstream.chat.android.offline.plugin.logic.channel.thread.internal
 
+import io.getstream.chat.android.client.extensions.internal.NEVER
 import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.offline.plugin.state.channel.thread.internal.ThreadMutableState
+import io.getstream.chat.android.offline.plugin.state.channel.thread.internal.toMutableState
 
 /**
  * The logic of the state of a thread. This class contains the logic of how to
  * update the state of the thread in the SDK.
+ *
+ * @property mutableState [ThreadMutableState]
  */
-internal interface ThreadStateLogic {
+internal class ThreadStateLogic(private val mutableState: ThreadMutableState) {
 
     /**
      * Return [ThreadMutableState] representing the state of the thread. Use this when you would like to
      * keep track of the state and would like to write a new state too.
      */
-    fun writeThreadState(): ThreadMutableState
+    fun writeThreadState(): ThreadMutableState {
+        return mutableState.toMutableState()
+    }
 
     /**
      * Deletes a message for the thread
      *
      * @param message [Message]
      */
-    fun deleteMessage(message: Message)
+    fun deleteMessage(message: Message) {
+        mutableState.deleteMessage(message)
+    }
 
     /**
      * Upsert message in the thread.
      *
      * @param message The message to be added or updated.
      */
-    fun upsertMessage(message: Message)
+    fun upsertMessage(message: Message) = upsertMessages(listOf(message))
 
     /**
      * Upsert messages in the channel.
@@ -51,5 +60,32 @@ internal interface ThreadStateLogic {
      * @param messages the list of [Message] to be upserted
      * new messages should be kept.
      */
-    fun upsertMessages(messages: List<Message>)
+    fun upsertMessages(messages: List<Message>) {
+        val oldMessages = mutableState.rawMessage.value
+        mutableState.upsertMessages(
+            messages.filter { newMessage -> isMessageNewerThanCurrent(oldMessages[newMessage.id], newMessage) }
+        )
+    }
+
+    private fun isMessageNewerThanCurrent(currentMessage: Message?, newMessage: Message): Boolean {
+        return if (newMessage.syncStatus == SyncStatus.COMPLETED) {
+            (currentMessage?.lastUpdateTime() ?: NEVER.time) <= newMessage.lastUpdateTime()
+        } else {
+            (currentMessage?.lastLocalUpdateTime() ?: NEVER.time) <= newMessage.lastLocalUpdateTime()
+        }
+    }
+
+    private fun Message.lastUpdateTime(): Long = listOfNotNull(
+        createdAt,
+        updatedAt,
+        deletedAt,
+    ).maxOfOrNull { it.time }
+        ?: NEVER.time
+
+    private fun Message.lastLocalUpdateTime(): Long = listOfNotNull(
+        createdLocallyAt,
+        updatedLocallyAt,
+        deletedAt,
+    ).maxOfOrNull { it.time }
+        ?: NEVER.time
 }
